+++ components/swiper/index.vue
@@ -100,6 +100,7 @@ export default {
       dragging: false,
       userScrolling: null,
       isInitial: false,
+      duration: 0,
       index: 0, // real index (swiper perspective)
       fromIndex: 0, // display index (user perspective)
       toIndex: 0, // display index
@@ -113,10 +114,19 @@ export default {
       timer: null,
       noDrag: false,
       scroller: null,
-      isStoped: false,
+      isStoped: true,
       $swiper: null,
       transitionEndHandler: null,
     }
+  },
+
+  watch: {
+    autoplay: {
+      handler(val) {
+        this.duration = val
+      },
+      immediate: true,
+    },
   },
 
   computed: {
@@ -144,29 +154,23 @@ export default {
   beforeMount
   */
   mounted() {
-    this.ready = true
-    this.$swiper = this.$el.querySelector('.md-swiper-container')
-    this.$swiperBox = this.$el.querySelector('.md-swiper-box')
-    this.$nextTick(() => {
-      this.$_reInitItems()
-      this.$_startPlay()
-      window.addEventListener('resize', this.$_resize)
-    })
+    this.$_resizeEnterBehavior()
   },
   /*
   beforeUpdate
   updated
-  activated
-  deactivated
-  beforeDestroy
   */
+  activated() {
+    this.$_resizeEnterBehavior()
+  },
+  deactivated() {
+    this.$_resizeLeaveBehavior()
+  },
+  /**
+   beforeDestroy
+   */
   destroyed() {
-    this.ready = false
-    this.$_clearTimer()
-    window.removeEventListener('resize', this.$_resize)
-    if (this.__resizeTimeout__) {
-      clearTimeout(this.__resizeTimeout__)
-    }
+    this.$_resizeLeaveBehavior()
   },
   /*
   errorCaptured
@@ -179,8 +183,11 @@ export default {
       if (this.__resizeTimeout__) {
         clearTimeout(this.__resizeTimeout__)
       }
+
+      // if swiper stoped originally, keep status
+      const startIndex = this.index
       this.__resizeTimeout__ = setTimeout(() => {
-        this.$_reInitItems()
+        this.$_reInitItems(startIndex)
       }, 300)
     },
     $_onDragStart(e) {
@@ -201,6 +208,7 @@ export default {
       if (this.isPrevent) {
         e.preventDefault()
       }
+      /* istanbul ignore if */
       if (!this.dragging) {
         return
       }
@@ -210,11 +218,13 @@ export default {
       if (this.isPrevent) {
         e.preventDefault()
       }
+      /* istanbul ignore if */
       if (this.userScrolling) {
         this.dragging = false
         this.dragState = {}
         return
       }
+      /* istanbul ignore if */
       if (!this.dragging) {
         return
       }
@@ -293,6 +303,12 @@ export default {
     },
 
     $_opacity(animate = true, opacity) {
+      const children = this.$children
+      /* istanbul ignore if */
+      if (!children || !children.length) {
+        return
+      }
+      /* istanbul ignore if */
       if (typeof opacity !== 'undefined') {
         let toIndex = 0
         let fromIndex = this.toIndex
@@ -311,16 +327,16 @@ export default {
             toIndex = 0
           }
         }
-        const from = this.$children[fromIndex].$el
-        const to = this.$children[toIndex].$el
+        const from = children[fromIndex].$el
+        const to = children[toIndex].$el
         from.style.opacity = 1 - Math.abs(opacity)
         from.style.transition = animate ? 'opacity 300ms ease' : ''
         to.style.opacity = Math.abs(opacity)
         return
       }
 
-      const from = this.$children[this.fromIndex].$el
-      const to = this.$children[this.toIndex].$el
+      const from = children[this.fromIndex].$el
+      const to = children[this.toIndex].$el
       from.style.opacity = 0
       from.style.transition = animate ? 'opacity 500ms ease' : ''
       to.style.opacity = 1
@@ -331,18 +347,26 @@ export default {
       }
     },
 
-    $_initState(children) {
+    $_initState(children, startIndex) {
       this.oItemCount = children.length
       this.rItemCount = children.length
       this.noDrag = children.length === 1 || !this.dragable
-      this.index = this.defaultIndex >= 0 && this.defaultIndex < children.length ? parseInt(this.defaultIndex) : 0
+
+      this.index =
+        startIndex !== undefined
+          ? this.$_calcDisplayIndex(startIndex)
+          : this.defaultIndex >= 0 && this.defaultIndex < children.length ? parseInt(this.defaultIndex) : 0
+
       this.firstIndex = 0
       this.lastIndex = children.length - 1
-      this.fromIndex = this.index === this.firstIndex ? this.lastIndex : this.index + 1
+      this.fromIndex =
+        this.index === this.firstIndex
+          ? this.lastIndex
+          : this.index === this.lastIndex ? this.firstIndex : this.index + 1
       this.toIndex = this.index
     },
 
-    $_reInitItems() {
+    $_reInitItems(startIndex) {
       const children = this.$children
 
       if (!children || !children.length) {
@@ -350,8 +374,7 @@ export default {
       }
 
       this.$_getDimension()
-
-      this.$_initState(children)
+      this.$_initState(children, startIndex)
 
       if (this.isSlide) {
         this.$_backupItem(children)
@@ -364,16 +387,17 @@ export default {
     },
 
     $_startPlay() {
-      if (this.autoplay > 0 && this.oItemCount > 1) {
+      if (this.duration > 0 && this.oItemCount > 1) {
         this.$_clearTimer()
         this.timer = setInterval(() => {
+          /* istanbul ignore if */
           if (!this.isLoop && this.index >= this.rItemCount - 1) {
             return this.$_clearTimer()
           }
           if (!this.dragging) {
             this.next()
           }
-        }, this.autoplay)
+        }, this.duration)
       }
     },
 
@@ -392,6 +416,7 @@ export default {
         if ((!vertical && currentTop === startTop) || (vertical && currentLeft === startLeft)) {
           return false
         } else {
+          /* istanbul ignore next */
           if (diffX * diffX + diffY * diffY >= 25) {
             const _touchAngle = Math.atan2(Math.abs(diffY), Math.abs(diffX)) * 180 / Math.PI
             return !vertical ? _touchAngle > this.touchAngle : 90 - _touchAngle > this.touchAngle
@@ -440,6 +465,7 @@ export default {
       if (!towards) {
         return
       }
+      /* istanbul ignore next */
       if (options && options.index !== undefined) {
         this.index = options.index
       } else if (towards === 'prev') {
@@ -476,7 +502,7 @@ export default {
       setTimeout(() => {
         const isFirstItem = this.isFirstItem && this.isLoop
         const isLastItem = this.isLastItem && this.isLoop
-
+        /* istanbul ignore next */
         this.transitionEndHandler = () => {
           // Recover first and last page
           if (isLastItem) {
@@ -538,7 +564,7 @@ export default {
       let offsetLeft = dragState.currentLeft - dragState.startLeft
       let offsetTop = dragState.currentTop - dragState.startTop
       this.userScrolling = this.$_isScroll(dragState, Math.abs(offsetLeft), Math.abs(offsetTop))
-
+      /* istanbul ignore if */
       if (this.userScrolling) {
         return
       }
@@ -577,7 +603,7 @@ export default {
       const isFastDrag = dragDuration < PAGING_DURATION
 
       if (isFastDrag && dragState.currentLeft === undefined) {
-        this.play(this.autoplay)
+        this.play(this.duration)
         return
       }
 
@@ -609,7 +635,25 @@ export default {
 
       this.dragState = {}
 
-      this.play(this.autoplay)
+      this.play(this.duration)
+    },
+    $_resizeEnterBehavior() {
+      this.ready = true
+      this.$swiper = this.$el.querySelector('.md-swiper-container')
+      this.$swiperBox = this.$el.querySelector('.md-swiper-box')
+      this.$nextTick(() => {
+        this.$_reInitItems()
+        this.play(this.duration)
+        window.addEventListener('resize', this.$_resize)
+      })
+    },
+    $_resizeLeaveBehavior() {
+      this.ready = false
+      this.$_clearTimer()
+      window.removeEventListener('resize', this.$_resize)
+      if (this.__resizeTimeout__) {
+        clearTimeout(this.__resizeTimeout__)
+      }
     },
 
     // MARK: events handler, 如 $_onButtonClick
@@ -637,19 +681,20 @@ export default {
       })
 
       // restart timer
-      this.play(this.autoplay)
+      this.play(this.duration)
     },
 
     getIndex() {
       return this.$_calcDisplayIndex(this.index)
     },
 
-    play(autoplay = 3000) {
+    play(duration = 3000) {
       this.$_clearTimer()
-      if (autoplay < 500) {
+      if (duration < 500) {
         return
       }
-      this.autoplay = autoplay || this.autoplay
+
+      this.duration = duration || this.autoplay
       this.$_startPlay()
       this.isStoped = false
     },
@@ -663,11 +708,12 @@ export default {
       if (!this.ready) {
         return
       }
+      /* istanbul ignore next */
       this.$nextTick(() => {
         this.$_clearTimer()
         this.$_reInitItems()
-        if (this.autoplay > 0 && !this.isStoped) {
-          this.$_startPlay()
+        if (!this.isStoped) {
+          this.play(this.duration)
         }
       })
     },
@@ -676,11 +722,12 @@ export default {
       if (!this.ready) {
         return
       }
+      /* istanbul ignore next */
       this.$nextTick(() => {
         this.$_clearTimer()
         this.$_reInitItems()
-        if (this.autoplay > 0 && !this.isStoped) {
-          this.$_startPlay()
+        if (!this.isStoped) {
+          this.play(this.duration)
         }
       })
     }, 50),